# openapi/panvex.yaml
...
openapi: 3.0.3
info:
  title: Panvex Control Plane API
  description: |
    Operator-facing HTTP API for the Panvex fleet management
    control-plane. Migrated incrementally from hand-rolled Zod
    schemas; consult `web/src/shared/api/schemas/` for endpoints not
    yet covered here.
  version: "1.0.0-pre"
  contact:
    name: Panvex maintainers
    url: https://github.com/lost-coder/panvex
servers:
  - url: /
    description: Same-origin (the panel serves both API and UI)
paths:
...
  /api/agents/{id}/update:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: dispatchAgentUpdate
      summary: Enqueue a self-update job for an agent
      description: |
        Schedules an `agent_self_update` job carrying the resolved
        binary asset metadata. The agent picks it up via its
        long-poll session.
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchAgentUpdateRequest'
      responses:
        '200':
          description: Update job dispatched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchAgentUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/agents/{id}/telemt/update:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: dispatchTelemtUpdate
      summary: Enqueue a telemt.update job for an agent's managed Telemt install
      description: |
        Schedules a `telemt.update` job that downloads, verifies, and swaps
        the agent's MANAGED Telemt binary in place (distinct from
        `POST /api/agents/{id}/update`, which replaces the AGENT's own
        binary), then runs a supervised restart per the agent's configured
        update strategy (`PUT /api/agents/{id}/telemt/update-strategy`).
        `version` defaults to the panel's cached latest known Telemt
        release when omitted. A guard chain (see `DispatchTelemtUpdateError`)
        rejects the request with 409 before enqueueing when the strategy is
        unconfigured, does not support a binary swap, the agent's live probe
        reports the update as unavailable, or no version is known — checked
        in that order. Admin-only, same trust bar as the update-strategy
        endpoints.
      tags: [agents]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchTelemtUpdateRequest'
      responses:
        '202':
          description: Update job dispatched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchTelemtUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            The update cannot be dispatched right now — see `code` in the
            response body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchTelemtUpdateError'
...
components:
  schemas:
...
    Error:
      type: object
      description: Standard error envelope used by every 4xx / 5xx response.
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable error code (e.g. `not_found`, `forbidden`).
        message:
          type: string
          description: Human-readable message; safe to display to operators.
        details:
          type: object
          description: Optional structured context. Shape varies per error code.
          additionalProperties: true
...
    DispatchTelemtUpdateRequest:
      type: object
      description: |
        Body of `POST /api/agents/{id}/telemt/update`. Both fields are
        optional — an empty body dispatches the panel's cached latest known
        Telemt release with no downgrade allowance.
      properties:
        version:
          type: string
          description: |
            Target Telemt version (bare semver, e.g. "3.4.25"). Omit to use
            the panel's cached latest known Telemt release
            (`telemt_latest_version` in the updates state); a 409
            `no_known_release` is returned if neither is available.
        allow_downgrade:
          type: boolean
          description: |
            Allow installing a version older than the one currently
            running (e.g. an emergency rollback). Defaults to false.
    DispatchTelemtUpdateResponse:
      type: object
      required: [job_id]
      properties:
        job_id:
          type: string
    DispatchTelemtUpdateError:
      type: object
      description: |
        409 error envelope for `POST /api/agents/{id}/telemt/update`: `code`
        is the machine-readable guard that rejected the request.
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable message; safe to display to operators.
        code:
          type: string
          enum:
            - strategy_not_configured
            - mode_not_binary
            - update_unavailable
            - no_known_release
          description: |
            strategy_not_configured: no update strategy has been PUT for
            this agent yet. mode_not_binary: the configured strategy's mode
            is "docker" or "none" — no in-place binary swap path exists.
            update_unavailable: the agent's live probe reports it cannot
            offer an in-place update right now. no_known_release: no
            version was given and the panel has never cached a latest
            Telemt release.
...
    DispatchAgentUpdateRequest:
      type: object
      required: [version]
      properties:
        version:
          type: string
          description: Target agent binary version.
    DispatchAgentUpdateResponse:
      type: object
      required: [job_id, status, version]
      properties:
        job_id:
          type: string
        status:
          type: string
          example: dispatched
        version:
          type: string
...
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Caller is not authenticated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Caller is authenticated but lacks the required role/scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Requested resource does not exist or is out-of-scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: panvex_session
      description: |
        Server-issued session cookie set by `POST /api/login`. Marked
        Secure + HttpOnly + SameSite=Lax in production; the bare
        non-prefixed name above is the dev fallback for plain-HTTP
        loopback (the production runtime sets `__Host-panvex_session`
        instead — same auth flow, different cookie attribute set).
security:
  - sessionCookie: []